[refactor](home): 홈 메인 추천 관련 컴포저블 recommendation/main 으로 이동 (#13)
