+++ app/src/main/java/com/kuit/ourmenu/ui/home/component/recommendation/main/HomeMainRecommendation.kt
@@ -1,0 +1,41 @@
+package com.kuit.ourmenu.ui.home.component.recommendation.main
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun HomeMainRecommendation() {
+    Column(
+        modifier = Modifier
+            .padding(top = 16.dp)
+    ) {
+        HomeMainRecommendationText(
+            modifier = Modifier
+                .padding(horizontal = 20.dp)
+                .width(278.dp)
+                .height(144.dp)
+        )
+
+        HomeMainRecommendationList(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 32.dp)
+                .height(244.dp)
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun MainRecommendationListPreview() {
+    HomeMainRecommendation()
+}
